minor refactoring and adjustments, updated pkgs
