+++ appinfo/info.xml
@@ -4,7 +4,7 @@
 	<name>Profile picker</name>
 	<summary>Profile smart picker and link preview</summary>
 	<description><![CDATA[This app adds ability to search for user profiles via smart picker and link previews for them.]]></description>
-	<version>0.2.2</version>
+	<version>0.2.3</version>
 	<licence>agpl</licence>
 	<author mail="andrey18106x@gmail.com" homepage="https://github.com/andrey18106">Andrey Borysenko</author>
 	<namespace>UsersPicker</namespace>
