+++ img/LICENSES.md
@@ -1,0 +1,8 @@
+# Licenses
+
+## profile.svg, profile-dark.svg
+
+* Created by: Google
+* License: Apache License version 2.0
+* Link: https://pictogrammers.com/library/mdi/icon/account/
+
