+++ lib/AppInfo/Application.php
@@ -1,27 +1,6 @@
 <?php
-/**
- * @copyright Copyright (c) 2023 Andrey Borysenko <andrey18106x@gmail.com>
- *
- * @copyright Copyright (c) 2023 Alexander Piskun <bigcat88@icloud.com>
- *
- * @author 2023 Andrey Borysenko <andrey18106x@gmail.com>
- *
- * @license AGPL-3.0-or-later
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
+
+declare(strict_types=1);
 
 namespace OCA\UsersPicker\AppInfo;
 
@@ -32,13 +11,8 @@ use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\Collaboration\Reference\RenderReferenceEvent;
 
 use OCA\UsersPicker\Listener\UsersPickerReferenceListener;
-use OCA\UsersPicker\Reference\UsersPickerReferenceProvider;
+use OCA\UsersPicker\Reference\ProfilePickerReferenceProvider;
 
-/**
- * Class Application
- *
- * @package OCA\UsersPicker\AppInfo
- */
 class Application extends App implements IBootstrap {
 	public const APP_ID = 'users_picker';
 
@@ -47,7 +21,7 @@ class Application extends App implements IBootstrap {
 	}
 
 	public function register(IRegistrationContext $context): void {
-		$context->registerReferenceProvider(UsersPickerReferenceProvider::class);
+		$context->registerReferenceProvider(ProfilePickerReferenceProvider::class);
 		$context->registerEventListener(RenderReferenceEvent::class, UsersPickerReferenceListener::class);
 	}
