+++ package.json
@@ -35,24 +35,25 @@
 		"extends @nextcloud/browserslist-config"
 	],
 	"dependencies": {
-		"@nextcloud/auth": "^2.0.0",
-		"@nextcloud/axios": "^2.3.0",
-		"@nextcloud/dialogs": "^4.0.1",
-		"@nextcloud/initial-state": "^2.0.0",
-		"@nextcloud/l10n": "^2.1.0",
-		"@nextcloud/moment": "^1.2.1",
-		"@nextcloud/router": "^2.1.1",
-		"@nextcloud/vue": "^7.10.0",
+		"@nextcloud/auth": "^2.2.1",
+		"@nextcloud/axios": "^2.4.0",
+		"@nextcloud/dialogs": "^5.0.3",
+		"@nextcloud/initial-state": "^2.1.0",
+		"@nextcloud/l10n": "^2.2.0",
+		"@nextcloud/moment": "^1.3.1",
+		"@nextcloud/router": "^2.2.0",
+		"@nextcloud/vue": "^8.4.0",
+		"debounce": "^2.0.0",
 		"vue": "^2.7.14",
 		"vue-click-outside": "^1.1.0",
 		"vue-material-design-icons": "^5.2.0"
 	},
 	"devDependencies": {
 		"@nextcloud/babel-config": "^1.0.0",
-		"@nextcloud/browserslist-config": "^2.3.0",
-		"@nextcloud/eslint-config": "^8.2.1",
+		"@nextcloud/browserslist-config": "^3.0.0",
+		"@nextcloud/eslint-config": "^8.3.0",
 		"@nextcloud/stylelint-config": "^2.3.0",
-		"@nextcloud/webpack-vue-config": "^5.5.0",
+		"@nextcloud/webpack-vue-config": "^6.0.0",
 		"eslint-webpack-plugin": "^4.0.1",
 		"stylelint-webpack-plugin": "^4.1.1"
 	},
