+++ src/reference.js
@@ -7,7 +7,7 @@ __webpack_public_path__ = linkTo('users_picker', 'js/') // eslint-disable-line
 
 registerWidget('users_picker_profile', async (el, { richObjectType, richObject, accessible }) => {
 	const { default: Vue } = await import(/* webpackChunkName: "reference-issue-lazy" */'vue')
-	const { default: UsersPickerReferenceWidget } = await import(/* webpackChunkName: "reference-issue-lazy" */'./views/UsersPickerReferenceWidget.vue')
+	const { default: UsersPickerReferenceWidget } = await import(/* webpackChunkName: "reference-issue-lazy" */'./views/ProfilePickerReferenceWidget.vue')
 	Vue.mixin({ methods: { t, n } })
 	const Widget = Vue.extend(UsersPickerReferenceWidget)
 	new Widget({
@@ -22,7 +22,7 @@ registerWidget('users_picker_profile', async (el, { richObjectType, richObject, 
 registerCustomPickerElement('profile_picker', async (el, { providerId, accessible }) => {
 	const { default: Vue } = await import(/* webpackChunkName: "vue-lazy" */'vue')
 	Vue.mixin({ methods: { t, n } })
-	const { default: UserProfilesCustomPicker } = await import(/* webpackChunkName: "image-picker-lazy" */'./components/UserProfilesCustomPicker.vue')
+	const { default: UserProfilesCustomPicker } = await import(/* webpackChunkName: "image-picker-lazy" */'./components/ProfilesCustomPicker.vue')
 	const Element = Vue.extend(UserProfilesCustomPicker)
 	const vueElement = new Element({
 		propsData: {
